outsourced check dirs and added it to gui
