+++ src/gui/homeWindow.py
@@ -10,6 +10,13 @@ from serverWindow import ServerWindow
 import yaml
 import os
 
+import sys
+from pathlib import Path
+current_dir = Path(__file__).parent
+project_root = current_dir.parent
+sys.path.append(str(project_root))
+from functions.checkDirs import checkDirs
+
 class HomeWindow(QMainWindow):
     def __init__(self):
         super().__init__()
@@ -77,6 +84,7 @@ class HomeWindow(QMainWindow):
                     self.ui.deviceSideContent.setFixedHeight(self.ui.deviceSideContent.height() + 75)
 
 if __name__ == "__main__":
+    checkDirs()
     app = QApplication([])
     homeWindow = HomeWindow()
     homeWindow.show()
